# backend/OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: Directory Wizard API
  description: |
    Multi-tenant Directory Creation Wizard API for creating and managing directory websites
    through a guided interface with integrated external service provisioning.
  version: 1.0.0
  contact:
    name: Directory Wizard Support
    email: support@directorywizard.com
    url: https://docs.directorywizard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.directorywizard.com/api
    description: Production server
  - url: https://staging-api.directorywizard.com/api
    description: Staging server
  - url: http://localhost:3001/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /listings/{id}/claim:
    get:
      summary: Get claim form data
      description: Get claim form data for a specific listing
      tags:
        - Listings
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Claim form data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimFormResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    post:
      summary: Submit claim
      description: Submit a new business claim
      tags:
        - Listings
      parameters:
        - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitClaimRequest'
      responses:
        '200':
          description: Claim submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitClaimResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ListingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^lst_[a-zA-Z0-9]+$'
      description: Listing ID
      example: lst_456789
  schemas:
...
    ClaimFormResponse:
      type: object
      properties:
        listing_id:
          type: string
          example: lst_456
        listing_title:
          type: string
          example: Best Pizza Place
        business_name:
          type: string
          example: Pizza Palace LLC
        current_info:
          $ref: '#/components/schemas/BusinessInfo'
        verification_methods:
          type: array
          items:
            $ref: '#/components/schemas/VerificationMethod'
        claim_window_days:
          type: integer
          example: 14
        terms_url:
          type: string
          format: uri
          example: https://example.com/terms/claims
    BusinessInfo:
      type: object
      properties:
        phone:
          type: string
          example: 555-0123
        email:
          type: string
          format: email
          example: info@pizza.com
        website:
          type: string
          format: uri
          example: https://pizza.com
        address:
          type: string
          example: 123 Main St, San Francisco, CA
    VerificationMethod:
      type: object
      properties:
        type:
          type: string
          enum: [EMAIL_DOMAIN, PHONE_NUMBER, BUSINESS_DOCUMENT, UTILITY_BILL]
          example: EMAIL_DOMAIN
        label:
          type: string
          example: Email Domain Verification
        description:
          type: string
          example: Verify ownership using business email domain
        requirements:
          type: array
          items:
            type: string
          example: ["Business email from domain pizza.com"]
    SubmitClaimRequest:
      type: object
      required:
        - verification_method
        - contact_email
        - contact_person
        - relationship
      properties:
        verification_method:
          type: string
          enum: [EMAIL_DOMAIN, PHONE_NUMBER, BUSINESS_DOCUMENT, UTILITY_BILL]
          example: EMAIL_DOMAIN
        contact_email:
          type: string
          format: email
          example: owner@pizza.com
        contact_person:
          type: string
          example: John Smith
        relationship:
          type: string
          enum: [owner, manager, employee, representative]
          example: owner
        additional_info:
          type: string
          example: I am the owner of this business and need to update the listing information.
    SubmitClaimResponse:
      type: object
      properties:
        claim_id:
          type: string
          example: clm_123456789
        listing_id:
          type: string
          example: lst_456
        status:
          type: string
          example: PENDING
        verification_method:
          type: string
          example: EMAIL_DOMAIN
        contact_email:
          type: string
          format: email
          example: owner@pizza.com
        expires_at:
          type: string
          format: date-time
        next_steps:
          type: array
          items:
            type: string
          example: ["Check your email for verification instructions", "Complete verification within 14 days"]
        estimated_review_time:
          type: string
          example: 2-3 business days
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        message:
          type: string
          example: Detailed error description
        field:
          type: string
          example: email
        code:
          type: string
          example: VALIDATION_ERROR
        timestamp:
          type: string
          format: date-time
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...